fix(symthaea): zero warnings, ignore slow tests, add from_genesis, harden unwraps
- Remove broken module declarations (coupled_physics.rs missing from repo)
- Add #[ignore] to 7 consciousness_integration tests that take >60s
- Add from_genesis() constructors to LearnableLTC and UnifiedLTC
- Replace partial_cmp().unwrap() with fallback ordering in physics modules
- Fix unnecessary parens warning in video source
- Replace bare unwrap with expect in trajectory trend()

All 57 consciousness_integration tests pass (7 ignored), 8 genesis
tests pass, full workspace compiles with only minor dead_code warnings.

+++ symthaea-hlb/src/perception/video/source.rs
@@ -415,7 +415,7 @@ impl MockVideoSource {
                 MockPattern::Burst { avg_gap_frames } => {
                     // Pseudo-random bursts using deterministic hash
                     let hash = seq.wrapping_mul(0x5851F42D4C957F2D).wrapping_add(0x14057B7EF767814F);
-                    let gap = (hash % (avg_gap_frames as u64 * 2)) as u64;
+                    let gap = hash % (avg_gap_frames as u64 * 2);
 
                     // Burst when sequence aligns with hash-based timing
                     if seq % (gap.max(1)) < 3 {
